Fix security vulnerabilities # 243
update webpack and webpack-dev-server. but this commit do not update webpack version 4 to 5.
after update to webpack5 failed test using `npm run test`.
this issue related with 'ENOENT'. looks like need to update `webpack.config.js`. but I thought `webpack` update issue is not included this issue. so not included.