Define JSONElement which represents the type exposed (#314)
`json` gives the impression that it can be used directly by the users. So
this commit changed the package that is not directly exposed to users to
`crdt` and rename `proxy` to `json`.

In addition, when the user uses custom methods on Array, it provides
types with getID to the user. Conversely, if the user uses native methods,
it returns types similar to the native interface as much as possible.