Improve creating `change` in deleteNodes (#316)
Previously, creating `change` was done for each node to delete. And
then if the `change` was continuous with the previous one, the two
`change`s were merged. This logic caused delays because it continuously
used `IndexOf` which can be maximum O(n) for each node.

To improve the delay, this commit changes to create only one `change`
for successive nodes from the beginning.