Apply range deletion in RGATreeSplit (#326)
From the stable version, range isolation in splay tree is applied.
By not applying node deletion yet, it is expected that removed
nodes return the correct index.

Using the splay operation, this commit separates nodes in
a given range as a subtree to delete. Since each node to delete
is not `splay`ed, performance improvement is expected.