Fix incorrect index for nodes newly created then concurrently removed (#334)
* Fix incorrect index for removed nodes

The already deleted nodes in the range to be deleted sometimes had
incorrect index since they were not included in either `nodesToKeep`
or `nodesToDelete`. This commit adds that missing index updates.

* Fix index bug for nodes newly created then removed

Nodes newly created and then concurrently removed in the range to be
deleted had incorrect `weight` since they were not included in either
`nodesToKeep` or `nodesToDelete`. This problem causes incorrect indexing,
and this commit fixes the missing weight updates.