Apply generics to support any types of properties in RichText (#354)
The attributes of rich text editors can be of various types.

This commit makes the export methods, RichText.edit and RichText.setStyle can
receive any type of value from the user. Using JSON.stringify and JSON.parse,
the values are treated as strings as before, so the other layers operate
unchanged.