add truncateNumber util function and replace Math.floor in Counter
Math.floor(n) returns integer n which is smaller or equal to n.
but in case of "Long" library, which used in js-sdk for purpose of handling large number, it only removes decimal point.
It works same way in Go sdk
Therefore, I try to replace Math.floor with remove decimal