Move Presence from Client to Document (#574)
To ensure atomic delivery of data for `Document` and `Presence` to users,
we are removing `Presence` from the `Client` and moving it to the
`Document`. In the future, we plan to further modify it by changing
`Document` to `Channel` or `Room` and rename `Root` as `Document`.

UpdatePresence

`UpdatePresence API` has been removed. Now, document and presence
changes are transmitted using the `PushPull API`. Consequently, the
presence is no longer transmitted as a component of the `Client`, but
rather included within the change. ChangePack now contains `presence`.

WatchDocument

In the `WatchDocument API`, both the request and response only include
the client ID, excluding the presence.

---------

Co-authored-by: Youngteac Hong <susukang98@gmail.com>