Improve tldraw example performance (#640)
In the tldraw example project, there's long latency when moving large
shapes (e.g., DrawShape). The reason is that the unchanged data
(e.g., drawing points in DrawShape) is sent when moving shapes.

So, in this PR, only point data (a point of shape on the whiteboard)
is sent to the Yorkie server.