Enhance Set and Add for representing nested elements (#691)
Previously, when setting values for objects or arrays, we followed a
process of creating outer shells first ({}) and then individually
setting inner values, resulting in multiple operations. This PR
introduces a modification to streamline this process by allowing the
direct setting of initial values in a single operation. This
enhancement simplifies the workflow and reduces the number of
generated operations.

With this PR, the set operation is enhanced to handle both outer
shells and inner values properly. This ensures that the set operation
is generated only once, and undo operations are applied correctly.

---------

Co-authored-by: Youngteac Hong <susukang98@gmail.com>