Handle escape string for strings containing quotes (#700)
In pull request #330, when implementing the escape string function, it
referred to the js-string-escape library. However, we missed a point
mentioned in its documentation:

> Note that the returned string is not necessarily valid JSON, since JSON
> disallows control characters, and \' is illegal in JSON.

This caused an error due to the mishandling of single quotes during the
escape string process. Since we don't need to handle single quotes
specially, we removed that part. Chromium implementation:

https://github.com/chromium/chromium/blob/fc0cb60674643692171a84bb94a92c8bf5781031/base/json/string_escape.cc#L33-L79

We also fixed a missing part in handling escape string for object keys.