Fix issue of incorrect display of remote selection in Quill example (#769)
This commit addresses the issue of the remote cursor not being
accurately displayed when local edits are made in the Quill example.
It ensures that the remote cursor is correctly rendered after each
edit.(Ref: y-quill)

It also adds extra checks to update the selection when there are
differences between the selection in Yorkie and Quill. For example,
if you add text before another user's selected text, Quill might
include the added text in the remote-selection. In such cases, it is
necessary to update the Quill selection in Yorkie.