Reverse TreeChanges when Deletion Operations in Tree (#774)
This commit addresses the issue where, in a tree data structure, when
deletion operations occur and are split into multiple events by
makeDeletionChanges, the order of events needs to be reversed. If you
don't reverse the order, you may lead to situations where adjustments
to the range are necessary for subscribers of the events.

More detailed information on this scenario can be found in issue #773.

When deletion operations result in multiple events due to the
makeDeletionChanges() function, in order to prevent issues for
subscribers handling these events, the order of events must be
reversed. By reversing the order of events, there is no need to adjust
the range of the next event by the previous event.