Fix gc for multiple nodes in text and tree type (#806)
There was an issue in the garbage collection process when dealing with
multiple nodes in text(tree) types. When some nodes in text were
collected, the corresponding text was removed from
elementHasRemovedNodesSetByCreatedAt in the root, preventing the GC from
correctly processing the remaining nodes.

This PR addresses the inconsistency in the GC process by modifying the
condition under which text is removed from
elementHasRemovedNodesSetByCreatedAt. Instead of removing text when some
nodes are garbage collected, text will now only be removed when all
nodes have been garbage collected. This change ensures that GC behaves
consistently even when multiple nodes are involved.