Implement RHT.GC (#815)
This commit adds GC to RHT data structure. RHT is used in Tree and
Text, and when Style methods are executed, values that need to be
GC'ed are added to the GC cache. This commit generalizes GC logic by
introducing abstract types GCParent and GCChild and managing them
flatly from the root.

---------

Co-authored-by: Youngteac Hong <susukang98@gmail.com>