Show removed node in devtools (#835)
This commit enhances Tree in DevTools with three main improvements.
Firstly, it introduces the ability to view removed nodes by
`Show removed nodes` button. Secondly, it adds a display of the
current revision number within the history tab, providing users with a
clear indication of the revision they are currently viewing. Lastly,
it displays node size in the node detail panel. This additional size
provides users with a more understanding of the node's characteristics.