Add taskQueue to handle each request one by one (#862)
When using JS SDK in environments like React, it is possible to assign
Client to React Components. However, when Component Unmount is
executed by moving location, resources like `yorkie.Client` need to be
`deactivated` or `yorkie.Document` need to be `detached`. The
challenge arises when it is not possible to await asynchronous
requests while deactivating one by one due to some reasons.

To address this issue, this commit introduces a task queue in `Client`
to handle all requests one by one, ensuring proper deactivation or
detachment.