Handle ErrClientNotActivated and ErrClientNotFound (#865)
This commit addresses detailed logic errors returned in RPC Error
response from the server, specifically `ErrClientNotActivated` and
`ErrClientNotFound`. It checks for these errors in the RPC response
and handles them by changing the Client's state.

In case of `ErrClientNotActivated` and `ErrClientNotFound`, Client's
state is changed to 'Deactivated,' and attached `Document`s are
changed to `Detached`. Therefore, if the application is subscribing
the status event using `Document.Subscribe`, it will receive `Detached`.

---------

Co-authored-by: blurfx <iam@xo.dev>