Fix non-deterministic test caused by realtime mode (#997)
In realtime, the client automatically calls sync() in the background
when changes occur. However, the test code also explicitly calls sync()
which can result in two sync() calls being made. This led to
non-deterministic test results, depending on the timing of the
background sync.

To fix this, document is now attached in manual, ensuring that sync is
only triggered by the test code.